Extract mcp routes helpers into submodule
Convert backend/mcp/routes.py to a package. The module-level surface
moves to two submodules:

- backend/mcp/routes/helpers.py (580 LOC): JSON-RPC envelopes, SSE
  streaming, prompt and resource definitions, request parsing, and the
  structured tool-call logger. Includes the long-running
  _stream_tool_call_response (~170 LOC inside the SSE generator).
- backend/mcp/routes/deps.py (18 LOC): the McpDeps dataclass, split out
  so helpers can import the type without circularity.

__init__.py drops from 986 to 445 LOC and continues to re-export McpDeps
and register_mcp_routes for app.py.

One SSE test (test_sse_tool_response_serializes_decimal_dates_and_datetimes)
patched backend.mcp.routes.call_tool to mock the tool invocation in the
SSE path; that call site now lives in backend.mcp.routes.helpers, so
the patch target updates accordingly. The four other call_tool patches
target the JSON path, which still imports call_tool in __init__.py, and
are unchanged.
